fixed form inputs and landing page
