+++ src/components/allegiances/MakeAllegiance.js
@@ -1,12 +1,14 @@
 import React from "react";
 import { useSelector, useDispatch } from "react-redux";
-import { Form } from "semantic-ui-react";
+import { Form, Icon, Modal, Segment } from "semantic-ui-react";
 
 import { axiosWithAuth } from "../utils/axiosWithAuth";
 import useGetToken from "../utils/useGetToken";
 import useForm from "../utils/useForm";
 import useImageUploader from "../utils/useImageUploader";
 import { ADD_ALLEGIANCE } from "../../reducers/userReducer";
+import styled from "styled-components"
+import Default from "../../assets/walter-avi.png"
 
 const MakeAllegiance = props => {
   //Fetches logged in user's info from redux store.
@@ -17,10 +19,10 @@ const MakeAllegiance = props => {
   const [token] = useGetToken();
 
   //Imports form custom hook to handle state, form entry and form submission.
-  const { values, handleChange, FormPage } = useForm(createAllegiance);
+  const { values, handleChange, handleSubmit, SubmitButton, ErrorMessage } = useForm(createAllegiance);
 
   //Imports image upload functions
-  const { image } = useImageUploader()
+  const { image, UploaderUI, modalOpen, setModal } = useImageUploader()
 
   const addAllegiance = async allegiance => {
     try {
@@ -60,33 +62,77 @@ const MakeAllegiance = props => {
     { text: "Other", value: "Other" }
   ]
 
-  //Inputs to add to the form.
-  const inputs = (
-    <>
-      <Form.Input
-        required
-        label="Allegiance Name"
-        placeholder="Allegiance Name"
-        onChange={handleChange}
-        value={values.allegiance_name || ""}
-        name="allegiance_name"
-        type="text"
-      />
-      <Form.Field
-        required
-        label="Sport"
-        placeholder="Sport"
-        onChange={handleChange}
-        value={values.sport || ""}
-        name="sport"
-        control="select"
-        type="text"
-        options={selectOptions}
-      />
-    </>
+  return (
+    <FormHolder>
+      <FormSegment raised color="violet" style={{ margin: 'auto' }}>
+        <Form onSubmit={handleSubmit} errors>
+          <BasicInfoHolder>
+            <UploadIcon name='edit' size='large' color='black' onClick={() => setModal(true)} />
+            <Modal
+              open={modalOpen}
+              onClose={() => setModal(false)}
+              trigger={<ProfilePic onClick={() => setModal(true)} src={image || values.image || Default} />}>
+              <UploaderUI displayImage={image || values.image} />
+            </Modal>
+          </BasicInfoHolder>
+          <Form.Input
+            required
+            label="Allegiance Name"
+            placeholder="Allegiance Name"
+            onChange={handleChange}
+            value={values.allegiance_name || ""}
+            name="allegiance_name"
+            type="text"
+          />
+          <Form.Field
+            required
+            label="Sport"
+            placeholder="Sport"
+            onChange={handleChange}
+            value={values.sport || ""}
+            name="sport"
+            control="select"
+            type="text"
+            options={selectOptions}
+          />
+          <ErrorMessage />
+          <SubmitButton />
+        </Form>
+      </FormSegment>
+    </FormHolder>
   )
-
-  return <FormPage inputs={inputs} />
 };
+const FormHolder = styled.div`
+background-color: #dee4e7;
+min-height: 90vh;
+padding-top: 5%;
+margin-top: -1.5%;
+@media (max-width: 320px) {
+	height: 87vh
+}`
 
+const FormSegment = styled(Segment)`
+width: 90%;
+margin: auto;
+marginBottom: 15%;
+`
+
+const UploadIcon = styled(Icon)`
+position: absolute;
+top: 2.8rem;
+left: 2.8rem`
+
+const ProfilePic = styled.img`
+border-color: black;
+object-fit: cover;
+width: 100px;
+height: 100px;
+border-radius: 50%;
+border: 1px solid black;
+flex: 0 0 auto;
+opacity: .6; `
+
+const BasicInfoHolder = styled.div`
+display: flex;
+flex-direction: row;`
 export default MakeAllegiance;
